feat: 유저 포인트 충전 통합 테스트

+++ src/main/java/com/hhplus/concertreservation/user/presentation/dto/request/ChargeUserPointRequest.java
@@ -1,6 +1,6 @@
 package com.hhplus.concertreservation.user.presentation.dto.request;
 
 public record ChargeUserPointRequest(
-        int amount
+        long amount
 ) {
 }
